+++ R/utilities.R
@@ -1,10 +1,7 @@
 
-# Exported Utilities ------------------------------------------------------
 
 
-
-
-# Internal Utilities ------------------------------------------------------
+# Report Utilities ------------------------------------------------------
 
 
 
@@ -210,6 +207,55 @@ filenm <- function(path) {
 
 }
 
+
+
+has_report <- function(outreq) {
+
+  ret <- FALSE
+
+  if (!is.null(outreq)) {
+    for (itm in outreq) {
+
+      if (!is.null(itm$report)) {
+        if (itm$report == TRUE)
+          ret <- TRUE
+
+      }
+    }
+  }
+
+
+  return(ret)
+
+}
+
+get_report_name <- function(outreq) {
+
+  ret <- ""
+
+  if (!is.null(outreq)) {
+
+    nms <- names(outreq)
+    for (nm in nms) {
+
+      if (!is.null(outreq[[nm]]$report)) {
+        if (outreq[[nm]]$report == TRUE)
+          ret <- nm
+
+      }
+    }
+  }
+
+
+  return(ret)
+
+}
+
+
+# Option utilities --------------------------------------------------------
+
+
+
 #' @noRd
 option_true <- function(options, name, default = FALSE) {
 
@@ -308,46 +354,12 @@ get_name <- function(nm = NULL, var = NULL, bylbl = NULL) {
 }
 
 
-
-has_report <- function(outreq) {
-
-  ret <- FALSE
-
-  if (!is.null(outreq)) {
-    for (itm in outreq) {
-
-      if (!is.null(itm$report)) {
-        if (itm$report == TRUE)
-          ret <- TRUE
-
-      }
-    }
-  }
+# Other Utilities ---------------------------------------------------------
 
 
-  return(ret)
-
-}
-
-get_report_name <- function(outreq) {
-
-  ret <- ""
-
-  if (!is.null(outreq)) {
-
-    nms <- names(outreq)
-    for (nm in nms) {
-
-      if (!is.null(outreq[[nm]]$report)) {
-        if (outreq[[nm]]$report == TRUE)
-          ret <- nm
-
-      }
-    }
-  }
+stackds <- function(data) {
 
 
-  return(ret)
 
 }