+++ tests/testthat/test-transpose.R
@@ -148,7 +148,7 @@ test_that("transpose7: proc_means and proc_transpose.", {
 
   mres <- proc_means(score, var = c("Test1", "Test2", "Final"),
                     stats = c("n", "mean", "std", "median", "min", "max"),
-                    out = out(direction = "wide", type = FALSE,
+                    out = out(shape = "wide", type = FALSE,
                               freq = FALSE))
 
   res <- proc_transpose(mres, id = "VAR", name = "STAT")
@@ -231,7 +231,7 @@ test_that("transpose10: transpose by with two variables and v() function.", {
 
 test_that("transpose11: copy parameter works recycle bigger.", {
 
-  stats <- proc_means(datm, out = out(direction = "wide",
+  stats <- proc_means(datm, out = out(shape = "wide",
                                       type = FALSE, freq = FALSE))
 
 
@@ -254,7 +254,7 @@ test_that("transpose11: copy parameter works recycle bigger.", {
 test_that("transpose12: copy parameter works recycle smaller", {
 
   stats <- proc_means(datm, stats = c("n", "mean", "median"),
-                      out = out(direction = "wide", type = FALSE, freq = FALSE))
+                      out = out(shape = "wide", type = FALSE, freq = FALSE))
 
 
   res1 <- data.frame(Group = "Group1", stats)
@@ -278,7 +278,7 @@ test_that("transpose13: copy parameter works recycle smaller", {
 
   mns <- proc_means(datm, stats = c("n", "mean", "median", "min", "max"),
                     var = "Age", class = "Flavor",
-                    out = out(direction = "wide", type = FALSE, freq = FALSE))
+                    out = out(shape = "wide", type = FALSE, freq = FALSE))
 
   mns[1, 1] <- "Total"
   mns
@@ -302,13 +302,13 @@ test_that("transpose14: copy parameter works with by groups", {
   stats <- proc_means(datm, stats = c("n", "mean", "median"),
                       var = c("Age", "PresentScore", "TasteScore"),
                       by = "Layers", out = out(type = FALSE, freq = FALSE,
-                                               direction ="wide"))
+                                               shape ="wide"))
 
   stats
 
   res1 <- data.frame(Group = "Group1", stats)
 
-  res2 <- proc_transpose(res1, copy = "Group", by = "Layers",
+  res2 <- proc_transpose(res1, copy = "Group", by = "BY",
                          name = "STAT", id = "VAR")
 
   res2
@@ -331,7 +331,7 @@ test_that("transpose15: all vars eliminates by, copy, and id from transpose", {
 
   res1 <- data.frame(Group = "Group1", stats)
 
-  res2 <- proc_transpose(res1, copy = "Group", by = "Layers",
+  res2 <- proc_transpose(res1, copy = "Group", by = "BY",
                          name = "STAT", id = "VAR")
 
   res2
@@ -372,7 +372,7 @@ test_that("transpose15: NSE works on transpose", {
 
   res1 <- data.frame(Group = "Group1", stats)
 
-  res2 <- proc_transpose(res1, copy = Group, by = Layers,
+  res2 <- proc_transpose(res1, copy = Group, by = BY,
                          name = STAT, id = VAR)
 
   res2